Add clear path feature and remove unused method
A new "clear path" feature has been added which allows users to clear the current path on the grid and start over. The button for this action has been added to the UI. The 'findPath' method in the PathFinder service, which was not being used, has been removed. Also, logging has been added to the main application file for better monitoring.